Source/Database_Service/requirements.txt
flask-sqlalchemy
psycopg2
requests

Flask>=0.12
flask-restful>=0.3.5
